Split types in dedicated modules
