+++ crates/client/rpc-core/src/lib.rs
@@ -7,389 +7,13 @@
 #[cfg(test)]
 mod tests;
 
-use std::collections::{BTreeMap, HashMap};
-
-use anyhow::Result;
-use base64::engine::general_purpose;
-use base64::Engine;
-use frame_support::storage::bounded_vec::BoundedVec;
-use hex::ToHex;
 use jsonrpsee::core::RpcResult;
 use jsonrpsee::proc_macros::rpc;
-use mp_starknet::execution::types::{ContractClassWrapper, EntryPointTypeWrapper, EntryPointWrapper, MaxEntryPoints};
-use serde::{Deserialize, Serialize};
 
-pub type FieldElement = String;
-pub type BlockNumber = u64;
-pub type BlockHash = FieldElement;
+pub mod types;
+pub mod utils;
 
-pub type ContractAddress = FieldElement;
-pub type ContractClassHash = FieldElement;
-
-/// A tag specifying a dynamic reference to a block
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub enum BlockTag {
-    /// The latest accepted block
-    #[serde(rename = "latest")]
-    Latest,
-    /// The current pending block
-    #[serde(rename = "pending")]
-    Pending,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct BlockHashAndNumber {
-    pub block_hash: FieldElement,
-    pub block_number: BlockNumber,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct SyncStatus {
-    pub starting_block_hash: FieldElement,
-    pub starting_block_num: BlockNumber,
-    pub current_block_hash: FieldElement,
-    pub current_block_num: BlockNumber,
-    pub highest_block_hash: FieldElement,
-    pub highest_block_num: BlockNumber,
-}
-
-/// Function call information
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct FunctionCall {
-    pub contract_address: FieldElement,
-    pub entry_point_selector: FieldElement,
-    pub calldata: Vec<FieldElement>,
-}
-
-// In order to mix tagged and untagged {de}serialization for BlockId (see starknet RPC standard)
-// in the same object, we need this kind of workaround with intermediate types
-mod block_id {
-    use super::*;
-
-    #[derive(Serialize, Deserialize, Clone)]
-    enum BlockIdTaggedVariants {
-        #[serde(rename = "block_hash")]
-        BlockHash(BlockHash),
-        #[serde(rename = "block_number")]
-        BlockNumber(BlockNumber),
-    }
-
-    #[derive(Serialize, Deserialize, Clone)]
-    #[serde(untagged)]
-    enum BlockIdUntagged {
-        Tagged(BlockIdTaggedVariants),
-        BlockTag(BlockTag),
-    }
-
-    /// A block hash, number or tag
-    #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-    #[serde(from = "BlockIdUntagged")]
-    #[serde(into = "BlockIdUntagged")]
-    pub enum BlockId {
-        BlockHash(BlockHash),
-        BlockNumber(BlockNumber),
-        BlockTag(BlockTag),
-    }
-
-    impl From<BlockIdUntagged> for BlockId {
-        fn from(value: BlockIdUntagged) -> Self {
-            match value {
-                BlockIdUntagged::Tagged(v) => match v {
-                    BlockIdTaggedVariants::BlockHash(h) => Self::BlockHash(h),
-                    BlockIdTaggedVariants::BlockNumber(n) => Self::BlockNumber(n),
-                },
-                BlockIdUntagged::BlockTag(t) => Self::BlockTag(t),
-            }
-        }
-    }
-
-    impl From<BlockId> for BlockIdUntagged {
-        fn from(value: BlockId) -> Self {
-            match value {
-                BlockId::BlockHash(h) => Self::Tagged(BlockIdTaggedVariants::BlockHash(h)),
-                BlockId::BlockNumber(n) => Self::Tagged(BlockIdTaggedVariants::BlockNumber(n)),
-                BlockId::BlockTag(t) => Self::BlockTag(t),
-            }
-        }
-    }
-}
-
-pub type Program = String;
-
-pub type Offset = String;
-pub type Selector = FieldElement;
-/// Deprecated Cairo entry point (pre Sierra)
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct DeprecatedCairoEntryPoint {
-    pub offset: Offset,
-    pub selector: Selector,
-}
-
-pub type DeprecatedConstructor = Vec<DeprecatedCairoEntryPoint>;
-pub type DeprecatedExternal = Vec<DeprecatedCairoEntryPoint>;
-pub type DeprecatedL1Handler = Vec<DeprecatedCairoEntryPoint>;
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct DeprecatedEntryPointsByType {
-    #[serde(rename = "CONSTRUCTOR")]
-    pub constructor: DeprecatedConstructor,
-    #[serde(rename = "EXTERNAL")]
-    pub external: DeprecatedExternal,
-    #[serde(rename = "L1_HANDLER")]
-    pub l_1_handler: DeprecatedL1Handler,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
-pub enum FunctionABIType {
-    #[serde(rename = "function")]
-    #[default]
-    Function,
-    #[serde(rename = "l1_handler")]
-    LOneHandler,
-    #[serde(rename = "constructor")]
-    Constructor,
-}
-
-pub type FunctionName = String;
-pub type ParameterName = String;
-pub type ParameterType = String;
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct TypedParameter {
-    pub name: ParameterName,
-    #[serde(rename = "type")]
-    pub _type: ParameterType,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct FunctionABIEntry {
-    #[serde(rename = "type")]
-    pub _type: FunctionABIType,
-    pub name: FunctionName,
-    pub inputs: TypedParameter,
-    pub outputs: TypedParameter,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
-pub enum EventABIType {
-    #[serde(rename = "event")]
-    #[default]
-    Event,
-}
-
-pub type EventName = String;
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct EventABIEntry {
-    #[serde(rename = "type")]
-    pub _type: EventABIType,
-    pub name: EventName,
-    pub keys: TypedParameter,
-    pub data: TypedParameter,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
-pub enum StructABIType {
-    #[serde(rename = "struct")]
-    #[default]
-    Struct,
-}
-
-pub type StructName = String;
-pub type Size = i64;
-pub type StructMember = HashMap<String, serde_json::Value>;
-pub type Members = Vec<StructMember>;
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct StructABIEntry {
-    #[serde(rename = "type")]
-    pub _type: StructABIType,
-    pub name: StructName,
-    pub size: Size,
-    pub members: Members,
-}
-
-/// The resulting block information with transaction hashes
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum MaybePendingBlockWithTxHashes {
-    Block(BlockWithTxHashes),
-    PendingBlock(PendingBlockWithTxHashes),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BlockWithTxHashes {
-    pub status: BlockStatus,
-    pub block_hash: FieldElement,
-    /// The hash of this block's parent
-    pub parent_hash: FieldElement,
-    /// The block number (its height)
-    pub block_number: u64,
-    /// The new global state root
-    pub new_root: FieldElement,
-    /// The time in which the block was created, encoded in Unix time
-    pub timestamp: u64,
-    /// The Starknet identity of the sequencer submitting this block
-    pub sequencer_address: FieldElement,
-    /// The hashes of the transactions included in this block
-    pub transactions: Vec<FieldElement>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PendingBlockWithTxHashes {
-    /// The hashes of the transactions included in this block
-    pub transactions: Vec<FieldElement>,
-    /// The time in which the block was created, encoded in Unix time
-    pub timestamp: u64,
-    /// The Starknet identity of the sequencer submitting this block
-    pub sequencer_address: FieldElement,
-    /// The hash of this block's parent
-    pub parent_hash: FieldElement,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub enum BlockStatus {
-    #[serde(rename = "PENDING")]
-    Pending,
-    #[serde(rename = "ACCEPTED_ON_L2")]
-    AcceptedOnL2,
-    #[serde(rename = "ACCEPTED_ON_L1")]
-    AcceptedOnL1,
-    #[serde(rename = "REJECTED")]
-    Rejected,
-}
-
-/// A block hash, number or tag
-/// TODO: fix block_tag
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-#[serde(untagged)]
-pub enum ContractABIEntry {
-    FunctionABIEntry(FunctionABIEntry),
-    EventABIEntry(EventABIEntry),
-    StructABIEntry(StructABIEntry),
-}
-
-pub type ContractABI = Vec<ContractABIEntry>;
-/// Deprecated Cairo contract class (pre Sierra)
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct DeprecatedContractClass {
-    pub program: Program,
-    pub entry_points_by_type: DeprecatedEntryPointsByType,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub abi: Option<ContractABI>,
-}
-
-pub type SierraProgram = Vec<FieldElement>;
-pub type ContractClassVersion = String;
-
-pub type FunctionIndex = i64;
-/// Cairo sierra entry point
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct SierraEntryPoint {
-    pub selector: Selector,
-    pub function_idx: FunctionIndex,
-}
-
-pub type Constructor = Vec<SierraEntryPoint>;
-pub type External = Vec<SierraEntryPoint>;
-pub type L1Handler = Vec<SierraEntryPoint>;
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct EntryPointsByType {
-    #[serde(rename = "CONSTRUCTOR")]
-    pub constructor: Constructor,
-    #[serde(rename = "EXTERNAL")]
-    pub external: External,
-    #[serde(rename = "L1_HANDLER")]
-    pub l_1_handler: L1Handler,
-}
-pub type ABI = String;
-/// Cairo sierra contract class
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
-pub struct SierraContractClass {
-    pub sierra_program: SierraProgram,
-    pub contract_class_version: ContractClassVersion,
-    pub entry_points_by_type: EntryPointsByType,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub abi: Option<ABI>,
-}
-
-/// Starknet contract class
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-#[serde(untagged)]
-pub enum RPCContractClass {
-    DeprecatedContractClass(DeprecatedContractClass),
-    ContractClass(SierraContractClass),
-}
-
-/// Returns a `ContractClassWrapper` from a `RPCContractClass`
-pub fn to_rpc_contract_class(_contract_class_wrapped: ContractClassWrapper) -> Result<RPCContractClass> {
-    Ok(RPCContractClass::ContractClass(Default::default()))
-}
-
-/// Returns a base64 encoded and compressed string of the input bytes
-fn _compress_and_encode_base64(data: &[u8]) -> Result<String> {
-    let data_compressed = _compress(data)?;
-    Ok(_encode_base64(&data_compressed))
-}
-
-/// Returns a compressed vector of bytes
-fn _compress(data: &[u8]) -> Result<Vec<u8>> {
-    let mut gzip_encoder = flate2::write::GzEncoder::new(Vec::new(), flate2::Compression::fast());
-    serde_json::to_writer(&mut gzip_encoder, data)?;
-    Ok(gzip_encoder.finish()?)
-}
-
-/// Returns a base64 encoded string of the input bytes
-fn _encode_base64(data: &[u8]) -> String {
-    general_purpose::STANDARD.encode(data)
-}
-
-/// Returns a deprecated entry point by type from hash map of entry point types to entrypoint
-fn _to_deprecated_entrypoint_by_type(
-    entries: BTreeMap<EntryPointTypeWrapper, BoundedVec<EntryPointWrapper, MaxEntryPoints>>,
-) -> DeprecatedEntryPointsByType {
-    let mut constructor = vec![];
-    let mut external = vec![];
-    let mut l_1_handler = vec![];
-    entries.into_iter().for_each(|(entry_type, entry_points)| match entry_type {
-        EntryPointTypeWrapper::Constructor => {
-            constructor = entry_points.into_iter().map(Into::into).collect();
-        }
-        EntryPointTypeWrapper::External => {
-            external = entry_points.into_iter().map(Into::into).collect();
-        }
-        EntryPointTypeWrapper::L1Handler => {
-            l_1_handler = entry_points.into_iter().map(Into::into).collect();
-        }
-    });
-    DeprecatedEntryPointsByType { constructor, external, l_1_handler }
-}
-
-impl From<EntryPointWrapper> for DeprecatedCairoEntryPoint {
-    fn from(value: EntryPointWrapper) -> Self {
-        let selector: String = value.selector.encode_hex();
-        let selector = add_prefix(&selector);
-        let offset: String = value.offset.to_be_bytes().as_slice().encode_hex();
-        let offset = add_prefix(remove_leading_zeros(&offset));
-        Self { selector, offset }
-    }
-}
-
-/// Add 0x prefix to hex string
-fn add_prefix(s: &str) -> String {
-    format!("0x{}", s)
-}
-
-/// Remove leading zeros from str
-fn remove_leading_zeros(s: &str) -> &str {
-    s.trim_start_matches('0')
-}
-
-pub use block_id::BlockId;
-
-/// Boolean or SyncStatus
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub enum Syncing {
-    #[serde(rename = "sync_status")]
-    False(bool),
-    #[serde(rename = "sync_status")]
-    SyncStatus(SyncStatus),
-}
+use types::*;
 
 /// Starknet rpc interface.
 #[rpc(server, namespace = "starknet")]
