+++ crates/client/rpc-core/src/types/block/block_hash_and_number.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+use super::{BlockNumber, FieldElement};
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
+pub struct BlockHashAndNumber {
+    pub block_hash: FieldElement,
+    pub block_number: BlockNumber,
+}
