+++ crates/client/rpc-core/src/types/function_abi.rs
@@ -1,0 +1,23 @@
+use serde::{Deserialize, Serialize};
+
+use super::{FunctionName, TypedParameter};
+
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
+pub enum FunctionABIType {
+    #[serde(rename = "function")]
+    #[default]
+    Function,
+    #[serde(rename = "l1_handler")]
+    LOneHandler,
+    #[serde(rename = "constructor")]
+    Constructor,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
+pub struct FunctionABIEntry {
+    #[serde(rename = "type")]
+    pub _type: FunctionABIType,
+    pub name: FunctionName,
+    pub inputs: TypedParameter,
+    pub outputs: TypedParameter,
+}
