+++ crates/client/rpc-core/src/types/function_call.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+use super::FieldElement;
+
+/// Function call information
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default)]
+pub struct FunctionCall {
+    pub contract_address: FieldElement,
+    pub entry_point_selector: FieldElement,
+    pub calldata: Vec<FieldElement>,
+}
