+++ crates/client/rpc-core/src/types/rpc_contract_class.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+use super::{DeprecatedContractClass, SierraContractClass};
+
+/// Starknet contract class
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
+#[serde(untagged)]
+pub enum RPCContractClass {
+    DeprecatedContractClass(DeprecatedContractClass),
+    ContractClass(SierraContractClass),
+}
